Make useRef the same as in original code
